Start of the TITH (This Isn't That Hard) project
TITH aims to be a new, modern, complete FTN stack, focusing on being
usable with Synchronet. The code will be C11 only as much as possible,
with OS-specific code only used where absolutely necessary.

TITH will be leaving behind historical FidoNet baggage wherever
reasonable while still adhering to the original spirit of as many
documented interfaces as possible. As an example, the standard TIC
format *requires* and 8.3 filename for both the TIC file itself and
the file it describes. The TITH stack will instead have a very similar
control file (likely tac) that will be mostly the same, but with the
File line allowing long, UTF-8 filenames. Converting from TIC will
be handled by the TITH stack, but converting to TIC is someone else's
problem.

TITH will be fully 5D, without the 8-character limit on domains. The
current plan is to use FSC-0067 style node addresses (ie: fidonet#1:103/1
instead of 1:103/1@fidonet).

TITH will be UTF-8. It will aggressively convert things to UTF-8 and
stubbornly keep them that way.

TITH will be multiple small programs that each do one job. If there's
a documented interface for programs that do the same job, TITH will try
to use that interface.

TITH does not care if unmaintained legacy software does not work
properly with the netmail, files, or echomail that TITH produces.
As an example, the SPTH control paragraph from FSC-0067 is planned to
be implemented. The stripping of PATH paragraphs and pruning of
SEEN-BY paragraphs is almost certain to cause issues with ARCmail 0.6.
I don't care. I only care about things that are on FidoNet (and other
FTN networks) today.

TITH formats, protocols, and interfaces will be publicly documented
and each time one is completed or updated, an FTA-1002 compliant version
will be posted on the internet, and a post will be made in the FTSC
public echo with a summary of the document/changes, and the URL it's
hosted at. This will be the extent of the engagement with the FTSC.

I plan to work with my NC closely, and work through them to keep TITH
accepted by the FidoNet community. Should some things be rejected by
my NC (or *Cs above them), TITH will attempt to provide the service
using alternative means. For example, TITH will require a 43-character
public key be added to the nodelist. The nodelist line length limit
is currently 157 characters. If the *Cs don't approve of using 27% of
the possible line for a something only a small number of systems use
(which would be a reasonable decision ignoring the reasonableness of
157 characters as a limit), infrastructure to maintain a TITH-specific
nodelist format and distribute it to interested parties will be set
up with no complaints. The idea is to improve FidoNet by example, not
to start debates.

Initial development is around the mailer component, which will provide
FTS-5005 BinkleyTerm Style Outbound support and a slightly modified
FSP-1040 SRIF interface. Every time I consult that FSP, which has the
same number that was assigned to the last document I submitted to the
FTSC, I'll be reminded of why I don't plan to engage with the FTSC
regarding any of this.

TITHmailer (name subject to change, TMail?) will provide public key
authentication, with the public key in the nodelist. Transfers will
be encrypted.