Fix up alloc/free usage
Since this is being designed so it can be used by third-party software,
and since we no longer call exit() on errors, we need to actually
free all memory that we allocate... we can't just bail and let the
OS clean it up.

So, the current command and the config are now thread-local globals
which are automatically clean up on termination, and ephemeral allocs
now go onto a thread-local stack.

You need to pushAlloc() anything that's allocated (which checks if
the allocation failed), then popAlloc() when the value is free()d
or returned.

With that, we can clean up all the allocations and longjmp() out
on errors.