Add an initial bundle specification
Currently, a bundle only contains files.  This is sufficient to
develop a working mailer for a Binkley style outbound, which only
transfers files.

A type has been reserved to hold a Message, which will seamlessly
allow moving away from Type 2x packets... but that type won't be
defined until the tosser/packer is developed.