Start updating to conform with the standards
Still a fair bit of work needed, this prototype combined too many
layers and included some dead-end ideas.

The *new* design will be to separate the actual exchange of files
(ie: mailer) from the outbound/inbound bundling/processing (ie:
bundler/unbundler).

The mailer will still need to handle FileRequests and call the SRIF
handler (really the SRIF handler wrapper since we won't be producing
SRIF compliant files, and will want the handler to generate bundles).

Still need to decide if that will just be a special mode of the bunder,
or if it will be a completely separate program.